Reports/requirements.txt
# Python dependencies for PDF report generation and automation

# PDF generation
weasyprint>=60.0
markdown>=3.5
Pygments>=2.17.0

# Google Drive upload
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Email functionality (built-in, no install needed)
# - smtplib
# - email
